+++ networking/eth2/src/integration-test/java/tech/pegasys/teku/networking/eth2/AbstractRpcMethodIntegrationTest.java
@@ -13,6 +13,8 @@
 
 package tech.pegasys.teku.networking.eth2;
 
+import static tech.pegasys.teku.util.config.Constants.MAX_CHUNK_SIZE;
+
 import org.junit.jupiter.api.AfterEach;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.eth2.peers.Eth2Peer;
@@ -31,6 +33,7 @@ public abstract class AbstractRpcMethodIntegrationTest {
       TestSpecFactory.createMinimalWithAltairForkEpoch(altairEpoch);
   protected final UInt64 altairSlot = altairEnabledSpec.computeStartSlotAtEpoch(altairEpoch);
   private final Eth2P2PNetworkFactory networkFactory = new Eth2P2PNetworkFactory();
+  private final RpcEncoding rpcEncoding = RpcEncoding.createSszSnappyEncoding(MAX_CHUNK_SIZE);
 
   @AfterEach
   public void tearDown() throws Exception {
@@ -91,7 +94,7 @@ public abstract class AbstractRpcMethodIntegrationTest {
       final Eth2P2PNetwork remotePeerNetwork =
           networkFactory
               .builder()
-              .rpcEncoding(RpcEncoding.SSZ_SNAPPY)
+              .rpcEncoding(rpcEncoding)
               .recentChainData(peerStorage.recentChainData())
               .historicalChainData(peerStorage.chainStorage())
               .spec(remoteSpec)
@@ -100,7 +103,7 @@ public abstract class AbstractRpcMethodIntegrationTest {
       final Eth2P2PNetwork localNetwork =
           networkFactory
               .builder()
-              .rpcEncoding(RpcEncoding.SSZ_SNAPPY)
+              .rpcEncoding(rpcEncoding)
               .peer(remotePeerNetwork)
               .recentChainData(localStorage.recentChainData())
               .historicalChainData(localStorage.chainStorage())