+++ networking/eth2/src/integration-test/java/tech/pegasys/teku/networking/eth2/PeerStatusIntegrationTest.java
@@ -15,6 +15,7 @@ package tech.pegasys.teku.networking.eth2;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.teku.infrastructure.async.Waiter.waitFor;
+import static tech.pegasys.teku.util.config.Constants.MAX_CHUNK_SIZE;
 
 import java.time.Duration;
 import java.util.List;
@@ -41,6 +42,7 @@ public class PeerStatusIntegrationTest {
   private static final List<BLSKeyPair> VALIDATOR_KEYS = BLSKeyGenerator.generateKeyPairs(1);
   private final Spec spec = TestSpecFactory.createMinimalPhase0();
   private final Eth2P2PNetworkFactory networkFactory = new Eth2P2PNetworkFactory();
+  private final RpcEncoding rpcEncoding = RpcEncoding.createSszSnappyEncoding(MAX_CHUNK_SIZE);
   private final RecentChainData recentChainData1 = MemoryOnlyRecentChainData.create(spec);
   private final BeaconChainUtil beaconChainUtil1 =
       BeaconChainUtil.create(spec, recentChainData1, VALIDATOR_KEYS);
@@ -57,7 +59,6 @@ public class PeerStatusIntegrationTest {
 
   @Test
   public void shouldExchangeStatusMessagesOnConnection() throws Exception {
-    final RpcEncoding encoding = RpcEncoding.SSZ_SNAPPY;
     final RecentChainData recentChainData2 = MemoryOnlyRecentChainData.create();
     BeaconChainUtil.create(recentChainData2, VALIDATOR_KEYS).initializeStorage();
 
@@ -65,14 +66,14 @@ public class PeerStatusIntegrationTest {
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .recentChainData(recentChainData1)
             .startNetwork();
     final Eth2P2PNetwork network2 =
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .recentChainData(recentChainData2)
             .startNetwork();
 
@@ -93,12 +94,11 @@ public class PeerStatusIntegrationTest {
 
   @Test
   public void shouldUpdatePeerStatus() throws Exception {
-    final RpcEncoding encoding = RpcEncoding.SSZ_SNAPPY;
     final Eth2P2PNetwork network1 =
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .recentChainData(recentChainData1)
             .startNetwork();
 
@@ -108,7 +108,7 @@ public class PeerStatusIntegrationTest {
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .recentChainData(recentChainData2)
             .peer(network1)
             .startNetwork();
@@ -127,13 +127,12 @@ public class PeerStatusIntegrationTest {
 
   @Test
   public void shouldUpdatePeerStatusPeriodically() throws Exception {
-    final RpcEncoding encoding = RpcEncoding.SSZ_SNAPPY;
     Duration statusUpdateInterval = Duration.ofSeconds(2);
     final Eth2P2PNetwork network1 =
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .eth2StatusUpdateInterval(statusUpdateInterval)
             .recentChainData(recentChainData1)
             .startNetwork();
@@ -144,7 +143,7 @@ public class PeerStatusIntegrationTest {
         networkFactory
             .builder()
             .spec(spec)
-            .rpcEncoding(encoding)
+            .rpcEncoding(rpcEncoding)
             .eth2StatusUpdateInterval(statusUpdateInterval)
             .recentChainData(recentChainData2)
             .peer(network1)