+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/subnets/AttestationSubnetSubscriptionsTest.java
@@ -22,6 +22,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static tech.pegasys.teku.infrastructure.async.SafeFutureAssert.assertThatSafeFuture;
+import static tech.pegasys.teku.util.config.Constants.GOSSIP_MAX_SIZE;
 
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
@@ -64,7 +65,8 @@ public class AttestationSubnetSubscriptionsTest {
             gossipEncoding,
             recentChainData,
             processor,
-            recentChainData.getCurrentForkInfo().orElseThrow());
+            recentChainData.getCurrentForkInfo().orElseThrow(),
+            GOSSIP_MAX_SIZE);
 
     when(gossipNetwork.subscribe(any(), any())).thenReturn(mock(TopicChannel.class));
   }