+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/core/RpcResponseEncoderTest.java
@@ -14,6 +14,7 @@
 package tech.pegasys.teku.networking.eth2.rpc.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static tech.pegasys.teku.util.config.Constants.MAX_CHUNK_SIZE;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
@@ -41,7 +42,7 @@ final class RpcResponseEncoderTest {
   private final RpcContextCodec<?, StatusMessage> contextCodec =
       RpcContextCodec.noop(StatusMessage.SSZ_SCHEMA);
   private final RpcResponseEncoder<StatusMessage, ?> responseEncoder =
-      new RpcResponseEncoder<>(RpcEncoding.SSZ_SNAPPY, contextCodec);
+      new RpcResponseEncoder<>(RpcEncoding.createSszSnappyEncoding(MAX_CHUNK_SIZE), contextCodec);
 
   @Test
   public void shouldEncodeSuccessfulResponse() {